// libs/repositories/src/TenderlyRepository/types.ts
export interface SimulationInput {
  input: string;
  from: string;
  to: string;
  value?: string;
  gas?: number;
  gas_price?: string;
}

export interface SimulationData {
  link: string;
  status: boolean;
  id: string;
}

export interface TenderlyBundleSimulationResponse {
  simulation_results: TenderlySimulation[];
}

// types were found in Uniswap repository
// https://github.com/Uniswap/governance-seatbelt/blob/e2c6a0b11d1660f3bd934dab0d9df3ca6f90a1a0/types.d.ts#L123

type StateObject = {
  balance?: string;
  code?: string;
  storage?: Record<string, string>;
};

type ContractObject = {
  contractName: string;
  source: string;
  sourcePath: string;
  compiler: {
    name: 'solc';
    version: string;
  };
  networks: Record<
    string,
    {
      events?: Record<string, string>;
      links?: Record<string, string>;
      address: string;
      transactionHash?: string;
    }
  >;
};

export type TenderlySimulatePayload = {
  network_id: string;
  block_number?: number;
  transaction_index?: number;
  from: string;
  to: string;
  input: string;
  gas: number;
  gas_price?: string;
  value?: string;
  simulation_type?: 'full' | 'quick';
  save?: boolean;
  save_if_fails?: boolean;
  state_objects?: Record<string, StateObject>;
  contracts?: ContractObject[];
  block_header?: {
    number?: string;
    timestamp?: string;
  };
  generate_access_list?: boolean;
};

// --- Tenderly types, Response ---
// NOTE: These type definitions were autogenerated using https://app.quicktype.io/, so are almost
// certainly not entirely accurate (and they have some interesting type names)

export interface TenderlySimulation {
  transaction: Transaction;
  simulation: Simulation;
  contracts: TenderlyContract[];
  generated_access_list: GeneratedAccessList[];
}

interface TenderlyContract {
  id: string;
  contract_id: string;
  balance: string;
  network_id: string;
  public: boolean;
  export: boolean;
  verified_by: string;
  verification_date: null;
  address: string;
  contract_name: string;
  ens_domain: null;
  type: string;
  evm_version: string;
  compiler_version: string;
  optimizations_used: boolean;
  optimization_runs: number;
  libraries: null;
  data: Data;
  creation_block: number;
  creation_tx: string;
  creator_address: string;
  created_at: Date;
  number_of_watches: null;
  language: string;
  in_project: boolean;
  number_of_files: number;
  standard?: string;
  standards?: string[];
  token_data?: TokenData;
}
// ...
interface GeneratedAccessList {
  address: string;
  storage_keys: string[];
}

interface Simulation {
  id: string;
  project_id: string;
  owner_id: string;
  network_id: string;
  block_number: number;
  transaction_index: number;
  from: string;
  to: string;
  input: string;
  gas: number;
  gas_price: string;
  value: string;
  method: string;
  status: boolean;
  access_list: null;
  queue_origin: string;
  created_at: Date;
}

interface ErrorInfo {
  error_message: string;
  address: string;
}

export interface SimulationError {
  error: {
    id: string;
    message: string;
    slug: string;
  };
}

interface Transaction {
  hash: string;
  block_hash: string;
  block_number: number;
  from: string;
  gas: number;
  gas_price: number;
  gas_fee_cap: number;
  gas_tip_cap: number;
  cumulative_gas_used: number;
  gas_used: number;
  effective_gas_price: number;
  input: string;
  nonce: number;
  to: string;
  index: number;
  error_message?: string;
  error_info?: ErrorInfo;
  value: string;
  access_list: null;
  status: boolean;
  addresses: string[];
  contract_ids: string[];
  network_id: string;
  function_selector: string;
  transaction_info: TransactionInfo;
  timestamp: Date;
  method: string;
  decoded_input: null;
  // Note: manually added (partial keys of `call_trace`)
  call_trace: Array<{
    error?: string;
    input: string;
  }>;
}